Switched to InstallSymlink for creating symlinks
OpenRAVE previously created symlinks with an install(CODE ...) directive
during the install step. Symlinks created in this way are not packaged
correctly when debhelper to build a Debian package. I switched to the
InstallSymlink macro, from [1], which properly handles this case. This
script is part of Bro [2] and, thus, is is covered by a 3-clause BSD
license [3].

[1] https://github.com/bro/cmake/blob/master/InstallSymlink.cmake
[2] http://www.bro.org/
[3] https://github.com/bro/bro/blob/master/COPYING