Added geometry groups to the KinBody XML
There is now an optional "group" attribute that can be added to <Geom>
elements. If this tag is specified, the geometry is added to a separate
geometry group in the LinkInfo struct. This geometry can be queried or
applied at runtime just like any other geometry group.

The geometry is created as usual if the "group" attribute is omitted, is empty,
or has the value "self" (the name of the default geometry group)..