Enable Doxygen tagfile generation
This commit instructs Doxygen to generate a tag file for OpenRAVE's documentation.  A tag file is an XML file which lists the files/classes/members documented by the project.  The intention with this commit is so that the generated tag file will be hosted alongside the core OpenRAVE documentation (e.g. at http://openrave.org/docs/latest_stable/coreapi.tag).