rplanners: fix mismatching function signature in declaration/definition
The function CheckRamp is declared with an optional parameter options.
However, the definition of the function does not have the parameter. Add
the parameter to the function definition to match declaration and
definition, and declare the parameter as unused.

Otherwise, this may lead to undefined symbol errors while loading the
plugin.