ikfast: move sympy-compatibility patches to separate file
The new file sympy_compat.py contains the patches to keep ikfast
backward compatible with older sympy versions.
The separate file makes it easier to manage and keep an overview
of the patches. It could also be used in other modules that use
sympy, or just be ditched at some point in the future when
ikfast becomes fully compatible with more recent sympy versions.