ikfast: use trigsimp with backward-compatibility, not upward-compatibility
We need to expand sin/cos sums after using trigsimp, which is done now
by calling the user-defined trigsimp_expanded method.
For sympy < 0.7.4 where trigsimp used less minimization, this step
is not necessary, hence a patch for backward-compatibility.

Ikfast is now free of upward-compatibility patches. It is up-to-date
with sympy-0.7.6 (at least until the next bug pops up) and the patches
are only for backward-compatiblity.