PGP Encryption
This commit introduces initial support for reading encrypted files fetched from remote locations.
It is based on libgpgme and libgpgmepp, official C and C++ interface to GPG.

The recipient secret key simply needs to be present on the system when an encrypted file is received.
Note that procedures such as finding and selecting keys is unnecessary. If the correct key exists,
the received data will be automatically decrypted into plaintext. If the correct key does not exist
then an error is logged.

The implementation is such that _ParseEncryptedDocument will recurse back into _ParseDocument
upon decryption. The URI of the file should therefore contain an appropriate extension for
OpenRAVE to guess file type. (Ideally the MIME type should be set, but we don't have this information.)

Signed-off-by: Tan Li Boon <liboon.tan@mujin.co.jp>