New implementation of encrypt and decrypt
This commit moves relevant code for PGP encryption and decryption into
jsoncommon.cpp, and also implements a streaming interface on top of GpgME.

Additionally, reading from a file, as well as writing to a file, memory,
and stream are now implemented.

User should provide the name of a secret key to use when encrypting a file,
via the attribute 'openravekey'. If a key name is not provided, then
OpenRAVE will use the first key available, or throw an exception if none
are available.

Signed-off-by: Tan Li Boon <liboon.tan@mujin.co.jp>