Cache absence of collision bodies
Previously, when an FCL collision manager was ensuring it had all
bodies, it would only cache the collision body pointers if the body
actually had any geometry that was enabled for collision.
This meant that if the environment contained bodies with large numbers
of non-colliding links, those bodies would _not_ get cached, and result
in a lot of overhead every time a collision check is run.

With this patch, cache state is updated for non-colliding objects as
well to avoid this extra work.